API-DB function separated in another file
